Projects/UCID-App/2025-08-uc-conversational-app-v1-v1/UC Conversational App V1/frontend/app.py
from flask import Flask, request, jsonify
from flask_cors import CORS
import os
from dotenv import load_dotenv

load_dotenv()

app = Flask(__name__)
CORS(app)

@app.route('/api/health', methods=['GET'])
def health_check():
    return jsonify({"status": "healthy", "service": "UC Conversational App"})

if __name__ == '__main__':
    app.run(host='0.0.0.0', port=5000, debug=True)
