Projects/UCID-App/2025-08-uc-conversational-app-v1-v1/UC Conversational App V1/frontend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Flask-SQLAlchemy==3.1.1
